% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippings.R
\name{all_my_clippings}
\alias{all_my_clippings}
\title{Extract and clean full Kindle clippings file}
\usage{
all_my_clippings(data = clippings_txt)
}
\arguments{
\item{data}{Download the "My_Clippings.txt" file from your Kindle and read into R as clippings_txt.}
}
\description{
This function loads the "My Clippings.txt" file generated from your Kindle highlights and creates a tidy data frame.
}
\examples{
clippings_txt <- readr::read_file(here::here("data/My Clippings.txt"))
all_my_clippings(data = clippings_txt)


}
