% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_updates.R
\name{kindle_updates_gs}
\alias{kindle_updates_gs}
\title{Pull new highlights and writes them to Google Sheet}
\usage{
kindle_updates_gs(gsworkbook, gsworksheet)
}
\arguments{
\item{gsworkbook}{Name of the Google Sheets workbook (file name) where your Kindle clippings are managed.}

\item{gsworksheet}{Name of the Google Sheets workbook (tab name) where your Kindle clippings are managed.}
}
\description{
Pull list of highlights not already in googlesheets file. To use if you have Google Sheets doc to store your Kindle highlights.
}
\examples{
clippings_txt <- readr::read_file(here::here("data/My Clippings.txt"))
kindle_updates_gs(gsworkbook = "My Workbook Name", gsworksheet = "My Worksheet Name")

}
