% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindleclipr-package.R
\docType{package}
\name{kindleclipr-package}
\alias{kindleclipr}
\alias{kindleclipr-package}
\title{kindleclipr: Manage your Kindle highlights with R}
\description{
kindleclipr formats the file generated from your Kindle highlights into a tidy data frame.
}
\author{
\strong{Maintainer}: Alexandra Mangold \email{alexandra.a.mangold@gmail.com}

}
\keyword{internal}
