% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clippings.R
\name{tidy_clippings}
\alias{tidy_clippings}
\title{Tidy Clippings}
\usage{
tidy_clippings(data)
}
\arguments{
\item{data}{Reference "My_Clippings.txt" file downloaded from Kindle and read into R.}
}
\description{
This function puts clippings file into a tidy format
}
